change: health context was changed to status context
