+++ src/adapters/providers/health/Health.provider.ts
@@ -1,9 +1,9 @@
-import { IHealthProvider } from "../../../application/modules/health/providerContracts/IHealth.provider";
+import { IStatusProvider } from "../../../application/modules/status/providerContracts/IStatus.provider";
 import AppSettings from "../../../application/shared/settings/AppSettings";
 import { NumberUtil } from "../../../domain/shared/utils/NumberUtil";
 import httpClient from "../../../infrastructure/httpClient/index";
 
-export class HealthProvider implements IHealthProvider {
+export class HealthProvider implements IStatusProvider {
   async get(context: string, date: string): Promise<string> {
     const doRequest = NumberUtil.isEvenNumber(new Date().getTime());
