--- src/application/modules/health/providerContracts/IHealth.provider.ts
@@ -1,3 +1,0 @@
-export interface IHealthProvider {
-  get(context: string, date: string): Promise<string>;
-}
