+++ src/application/modules/status/providerContracts/IStatus.provider.ts
@@ -1,0 +1,3 @@
+export interface IStatusProvider {
+  get(context: string, date: string): Promise<string>;
+}
