+++ src/infrastructure/app/AppWrapper.ts
@@ -1,6 +1,6 @@
 import infraContainer from "../container";
 infraContainer.load();
-import healthController from "../../adapters/controllers/health/Health.controller";
+import statusController from "../../adapters/controllers/status/Status.controller";
 import routeWhiteListMiddleware from "../middleware/authorization/whiteList";
 import AppSettings from "../../application/shared/settings/AppSettings";
 import Encryption from "../../application/shared/security/encryption";
@@ -62,7 +62,7 @@ export default class AppWrapper {
         this.app.use(AppSettings.ServerRoot, TypeParser.cast<Application>(controller.router));
         console.log(`${controller?.constructor?.name} was initialized`);
       });
-    this.app.use(TypeParser.cast<RequestHandler>(healthController.resourceNotFound));
+    this.app.use(TypeParser.cast<RequestHandler>(statusController.resourceNotFound));
     this.loadErrorHandler();
   }
 
@@ -87,7 +87,7 @@ export default class AppWrapper {
       this.app.use(AppSettings.ServerRoot, TypeParser.cast<Application>(controller.router));
       console.log(`${controller?.constructor?.name} was loaded`);
     }
-    this.app.use(TypeParser.cast<RequestHandler>(healthController.resourceNotFound));
+    this.app.use(TypeParser.cast<RequestHandler>(statusController.resourceNotFound));
     this.loadErrorHandler();
 
     return Promise.resolve();
